+++ lib/dm-pagination.rb
@@ -26,7 +26,10 @@ require 'dm-aggregates'
 require 'dm-pagination/version'
 require 'dm-pagination/pagination'
 require 'dm-pagination/pager'
+require 'dm-pagination/view_helper'
 
 DataMapper::Model.send :include, DataMapper::Pagination
 DataMapper::Collection.send :include, DataMapper::Pagination
 DataMapper::Query.send :include, DataMapper::Pagination
+
+ActionView::Base.send :include, DataMapper::Pagination::ViewHelper::Rails if defined?(ActionView::Base)