+++ History.md
@@ -1,3 +1,8 @@
+0.0.2 / 2009-10-07
+==================
+
+  * added view helper for Rails
+  * fixed Pager#to_html (needs to know the base path of the current URL)
 
 0.0.1 / YYYY-MM-DD
 ==================